Added smarter typing for return value of mount()
