--- src/components/component-proxy.ts
@@ -1,24 +1,0 @@
-import { ComponentClass } from '../api/component.js'
-import { CreateElement } from '../api/create-element.js'
-import { VNode } from '../api/vnode.js'
-
-export interface ComponentProxy {
-  props: Record<string, any>
-  [state: string]: any
-  $element: Element
-  $vnode: VNode
-  $includes: Map<string, ComponentClass>
-  render: Render
-  $update: () => void
-  beforeInit: () => void
-  beforeMount: () => void
-  afterMount: () => void
-  beforeRender: () => void
-  afterRender: () => void
-  beforeUnmount: () => void
-  afterUnmount: () => void
-}
-
-export interface Render {
-  (createElement: CreateElement): VNode
-}
